refactor notes internals, adopt %notes http create flow from notes-groups-integration
