+++ packages/app/ui/components/NotesChannel/index.ts
@@ -1,4 +1,5 @@
-export { NotesNativeChannel, NotesNoteDetail } from './NotesNativeChannel';
+export { NotesNativeChannel } from './NotesNativeChannel';
+export { NotesNoteDetail } from './NotesNoteDetail';
 export { NotesPostCollection } from './NotesPostCollection';
 export { NotesDraftInput } from './NotesDraftInput';
 export { NotesPostContent } from './NotesPostContent';
